+++ operator/internal/manifests/distributor_test.go
@@ -8,9 +8,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	policyv1 "k8s.io/api/policy/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	v1 "github.com/grafana/loki/operator/apis/config/v1"
 	lokiv1 "github.com/grafana/loki/operator/apis/loki/v1"
 	"github.com/grafana/loki/operator/internal/manifests"
 )
@@ -194,22 +192,4 @@ func TestNewDistributorDeployment_TopologySpreadConstraints(t *testing.T) {
 			require.Equal(t, tc.ExpectedTopologySpreadContraint, depl.Spec.Template.Spec.TopologySpreadConstraints)
 		})
 	}
-}
-
-func TestNewDistributorDeployment_Affinity(t *testing.T) {
-	manifests.TestAffinity(t, manifests.LabelDistributorComponent, func(cSpec *lokiv1.LokiComponentSpec, nAffinity bool) client.Object {
-		return manifests.NewDistributorDeployment(manifests.Options{
-			Name:      "abcd",
-			Namespace: "efgh",
-			Stack: lokiv1.LokiStackSpec{
-				StorageClassName: "standard",
-				Template: &lokiv1.LokiTemplateSpec{
-					Distributor: cSpec,
-				},
-			},
-			Gates: v1.FeatureGates{
-				DefaultNodeAffinity: nAffinity,
-			},
-		})
-	})
 }