+++ operator/internal/manifests/ingester_test.go
@@ -8,7 +8,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	policyv1 "k8s.io/api/policy/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	v1 "github.com/grafana/loki/operator/apis/config/v1"
 	lokiv1 "github.com/grafana/loki/operator/apis/loki/v1"
@@ -187,22 +186,4 @@ func TestNewIngesterStatefulSet_TopologySpreadConstraints(t *testing.T) {
 			},
 		},
 	}, ss.Spec.Template.Spec.TopologySpreadConstraints)
-}
-
-func TestNewIngesterStatefulSet_Affinity(t *testing.T) {
-	manifests.TestAffinity(t, manifests.LabelIngesterComponent, func(cSpec *lokiv1.LokiComponentSpec, nAffinity bool) client.Object {
-		return manifests.NewIngesterStatefulSet(manifests.Options{
-			Name:      "abcd",
-			Namespace: "efgh",
-			Stack: lokiv1.LokiStackSpec{
-				StorageClassName: "standard",
-				Template: &lokiv1.LokiTemplateSpec{
-					Ingester: cSpec,
-				},
-			},
-			Gates: v1.FeatureGates{
-				DefaultNodeAffinity: nAffinity,
-			},
-		})
-	})
 }