+++ operator/internal/manifests/query-frontend_test.go
@@ -7,9 +7,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	policyv1 "k8s.io/api/policy/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	v1 "github.com/grafana/loki/operator/apis/config/v1"
 	lokiv1 "github.com/grafana/loki/operator/apis/loki/v1"
 )
 
@@ -147,22 +145,4 @@ func TestNewQueryFrontendDeployment_TopologySpreadConstraints(t *testing.T) {
 			},
 		},
 	}, depl.Spec.Template.Spec.TopologySpreadConstraints)
-}
-
-func TestNewQuerierFrontendDeployment_Affinity(t *testing.T) {
-	TestAffinity(t, LabelQueryFrontendComponent, func(cSpec *lokiv1.LokiComponentSpec, nAffinity bool) client.Object {
-		return NewQueryFrontendDeployment(Options{
-			Name:      "abcd",
-			Namespace: "efgh",
-			Stack: lokiv1.LokiStackSpec{
-				StorageClassName: "standard",
-				Template: &lokiv1.LokiTemplateSpec{
-					QueryFrontend: cSpec,
-				},
-			},
-			Gates: v1.FeatureGates{
-				DefaultNodeAffinity: nAffinity,
-			},
-		})
-	})
 }